migrate: move MergeNBSettings and GitOperations into core types
- for dependency injection (next commit) so that core doesn't depend on app
- Defines both interfaces in packages/core/src/types.ts and adds a barrel export in packages/core/src/index.ts
- settings.ts re-exports from core.
